Make backgammon(7)'s sleeps interruptible.
backgammon sleeps at three places in the code: when a human player
can't move at all (two places, one if the player could have doubled and
one if not) and after the computer's first move when the computer moves
first.  These sleeps are long enough to be annoying, especially the
former.

With this change, typing any character during the sleep breaks out of
the sleep (and eats the typed character).