Give the lpt driver a `raw' mode, which treats it as 5 input and 12 output pins.
Specifically, the status port is polled once per hz and changes produce
readable bytes; writes must be even-length and alternate bytes are
shoved out the data (0xff bits) control (0x0f bits) ports, with no
particular timing promises.

Raw mode is indicated by the 0x100 bit in the device minor number.  It
is not compatible with any of the other flag bits in the minor number,
and it is not, of course, possible to have the port open both raw and
non-raw at the same time.

Writes are always nonblocking because there is nothing to block on.
poll() and nonblocking reads are supported.  SIGIO generation is not.