First steps towards SCM_MEMORY.
This version defines SCM_MEMORY and reworks unp_{in,ex}ternalize to
support it (and, incidentally, multiple SCM_RIGHTS messages).  But, in
this version, no VM is actually attached to the SCM_MEMORY messages and
they come through holding exactly what the sender put in them, with no
change in the receiver's VM mappings.  It is just the infrastructure
changes.